Exporting StartStruct to use with external prover.
